+++ x/cronos/types/interfaces.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/vm"
-	gravitytypes "github.com/peggyjv/gravity-bridge/module/x/gravity/types"
+	gravitytypes "github.com/peggyjv/gravity-bridge/module/v2/x/gravity/types"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 )
@@ -45,7 +45,7 @@ type AccountKeeper interface {
 
 // GravityKeeper defines the expected gravity keeper interface
 type GravityKeeper interface {
-	ERC20ToDenomLookup(ctx sdk.Context, tokenContract string) (bool, string)
+	ERC20ToDenomLookup(ctx sdk.Context, tokenContract common.Address) (bool, string)
 	GetParams(ctx sdk.Context) (params gravitytypes.Params)
 }